Add dashboard view and draft utilities; enhance DBUtils with new queries
