+++ test_debug.py
@@ -3,4 +3,4 @@ from db.db_utils import DBUtils
 
 db = DBUtils("assets/db/db_test.sqlite3")
 
-print(db.get_all_medocs_names_quantity_expiration_date())
+print(db.get_all_mouvement_facture())
